+++ R/gather_draws.R
@@ -10,13 +10,26 @@
 #' @importFrom dplyr bind_rows group_by_at
 #' @importFrom rlang enquos
 #' @export
-gather_draws = function(model, ..., regex = FALSE, sep = "[, ]", ndraws = NULL, seed = NULL, n) {
+gather_draws = function(
+  model,
+  ...,
+  regex = FALSE,
+  sep = "[, ]",
+  ndraws = NULL,
+  seed = NULL,
+  draw_indices = c(".chain", ".iteration", ".draw"),
+  n
+) {
   ndraws = .Deprecated_argument_alias(ndraws, n)
 
   draws = sample_draws_from_model_(model, ndraws, seed)
 
+  draw_indices = intersect(draw_indices, names(draws))
   tidysamples = lapply(enquos(...), function(variable_spec) {
-    gather_variables(spread_draws_(draws, variable_spec, regex = regex, sep = sep))
+    gather_variables(
+      spread_draws_(draws, variable_spec, regex = regex, sep = sep, draw_indices = draw_indices),
+      exclude = c(draw_indices, ".row")
+    )
   })
 
   #get the groups from all the samples --- when we bind them together,