+++ R/ungather_draws.R
@@ -14,7 +14,7 @@ globalVariables(c("..dimension_values"))
 ungather_draws = function(
   data, ..., variable = ".variable", value = ".value", draw_indices = c(".chain", ".iteration", ".draw"), drop_indices = FALSE
 ) {
-
+  draw_indices = intersect(draw_indices, names(data))
   variable_specs = enquos(...)
 
   if (length(variable_specs) == 0) {