+++ tests/testthat/test.gather_draws.R
@@ -7,8 +7,6 @@ library(dplyr)
 library(tidyr)
 
 
-
-
 test_that("regular expressions for parameter names work on non-indexed parameters", {
   data(RankCorr, package = "ggdist")
 
@@ -39,6 +37,34 @@ test_that("gather_draws works on a combination of 0 and 1-dimensional values (wi
   result = gather_draws(RankCorr, tau[i], typical_r)
 
   # grouped tibble equivalence is too persnickety now, so test equivalence + equal groups
+  expect_equivalent(result, ref)
+  expect_equal(group_vars(result), group_vars(ref))
+})
+
+test_that("draw_indices works", {
+  df = data.frame(
+    .chain = rep(1:4, each = 4),
+    .iteration = rep(1:4, 4),
+    .draw = 1:16,
+    .warmup = rep(c(TRUE, TRUE, FALSE, FALSE), 4),
+    `x[1]` = 2:17,
+    `x[2]` = 3:18,
+    check.names = FALSE
+  )
+
+  ref = tibble(
+    i = rep(1:2, each = 16),
+    .chain = rep(1:4, each = 4, times = 2),
+    .iteration = rep(1:4, 8),
+    .draw = rep(1:16, 2),
+    .warmup = rep(c(TRUE, TRUE, FALSE, FALSE), 8),
+    .variable = "x",
+    .value = c(2:17, 3:18)
+  ) %>%
+    group_by(i, .variable)
+
+  result = gather_draws(df, x[i], draw_indices = c(".chain", ".iteration", ".draw", ".warmup"))
+
   expect_equivalent(result, ref)
   expect_equal(group_vars(result), group_vars(ref))
 })