+++ NEWS.md
@@ -1,3 +1,5 @@
+# tidybayes (development version)
+
 # tidybayes 3.0.6
 
 Deprecations: